backend/requirements.txt
Django>=3.2.0,<3.3.0
djangorestframework>=3.12.0,<3.13.0
django-filter>=2.4.0,<2.5.0
pytest>=6.2.0,<6.3.0
pytest-django>=4.4.0,<4.5.0
psycopg2>=2.9.0,<2.10.0
Pillow>=8.3.0,<8.4.0
django-cors-headers>=3.7.0,<3.8.0
python-decouple>=3.4,<3.5

flake8>=3.9.0,<3.10.0